Revamped Project Pages
Created new project schema type to breakdown specific projects through dynamic pages on site.
